requirements-dev.txt
-r requirements.txt

black==24.4.2
jsonschema==4.23.0
pytest-asyncio==0.24.0
